Update the nlpjs model

+++ core/src/main/java/com/xatkit/core/recognition/nlpjs/model/Classification.java
@@ -5,6 +5,6 @@ import lombok.Data;
 @Data
 public class Classification {
     private String intent;
-    private Double score;
+    private Float score;
 
 }
